Fix setup.py dependency cycle
The __init__.py can't be imported if the dependencies of this package aren't
installed. But they can't be installed if the module cant load setup.py. So we
got a dependencie cycle. All this is related to the version number, which is
loaded from __init__.
Hence we now parse the file only for the version string. So we must not load
the imports in __init__